// src/enums/uam.ts
export enum BU_TYPE_ENUM {
  TMS = 'TMS',
  WMS = 'WMS',
  UAM = 'UAM',
  HR = 'HR',
}

export enum BU_TYPE_PRIMARY_COLOR {
  TMS = '#1AA391',
  WMS = '#1890FF',
  UAM = '#2F54EB',
  HR = '#FAAD14',
}

export enum REGION_ID_ENUM {
  Singapore = 10001,
  Philippines,
  Thailand,
  ChinaShenZhen,
  ChinaChengDu,
}
export enum REGION_NAME_ENUM {
  Singapore = 'Singapore',
  Philippines = 'Philippines',
  Thailand = 'Thailand',
  China = 'China',
}